feat(test): refactor improve test quality and change productId from Long to Integer;

+++ src/main/java/com/ecommerce/pricing/application/usecase/GetAllPricesByProductIdUseCaseImpl.java
@@ -16,7 +16,7 @@ public class GetAllPricesByProductIdUseCaseImpl implements GetAllPricesByProduct
 
   @TimedLog
   @Override
-  public Flux<Price> execute(Long productId) {
+  public Flux<Price> execute(Integer productId) {
     return priceRepositoryPort.findAllPricesByProductId(productId);
   }
 }
